relocated some files, added tests for preserveLogger and wrote some readmes
